feat: update patch file to include added features
A short (technical) description of what I did, for future me and anyone wanting to do any further updates: the reason why it wasn't possible before to use the features in certain sections was because the game uses a timer at 0x0a18 to determine for how long the player will not be able to enter anything except Start (pause). If 0x0a18 is anything but zero, no further processing is done. The way I solved it is by before this check, move the already polled input to address 0x040c and keep it there. Then, run the code as usual. This allows for the game to still run as intended: the player will still not be able to move during these sections but will be able to reset as desired. Also, a nice side effect (and something I wanted to implement anyways) is that when you press Start, you are taken immediately to the level select screen. This happens because the game uses 0x040c to determine whether or not to take the player to the level select screen. It is also used in the original game to determine whether or not the player goes to the next act/level when pausing and pressing A, but that doesn't really matter here since that logic is overriden. And last but not least, this line of code '3FFB82:  33F9  move.w  ($FFFF04A0), ($FFFF04A2)' remains unexplained to me. So to future me or anyone else reading that, no idea why it's there :p